requirements.txt
numpy==1.22.3
opencv_contrib_python==4.5.5.64
Pillow==9.1.0
PyMuPDF==1.19.6
pytesseract==0.3.9

WTForms~=3.0.1
Flask~=2.1.1
PyPDF2~=1.27.9